Move getDefaultDeviceName into the Platforms
So we can have a sensible device name on Electron
